+++ src/features/Conversation/Extras/Assistant.tsx
@@ -1,9 +1,7 @@
-import { ModelTag } from '@lobehub/icons';
 import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
-import { useAgentStore } from '@/store/agent';
-import { agentSelectors } from '@/store/agent/slices/chat';
+import { LOADING_FLAT } from '@/const/message';
 import { useChatStore } from '@/store/chat';
 import { chatSelectors } from '@/store/chat/selectors';
 import { ChatMessage } from '@/types/message';
@@ -12,34 +10,28 @@ import { RenderMessageExtra } from '../types';
 import ExtraContainer from './ExtraContainer';
 import TTS from './TTS';
 import Translate from './Translate';
+import Usage from './Usage';
 
 export const AssistantMessageExtra: RenderMessageExtra = memo<ChatMessage>(
-  ({ extra, id, content }) => {
-    const model = useAgentStore(agentSelectors.currentAgentModel);
+  ({ extra, id, content, metadata, tools }) => {
     const loading = useChatStore(chatSelectors.isMessageGenerating(id));
 
-    const showModelTag = extra?.fromModel && model !== extra?.fromModel;
-    const showTranslate = !!extra?.translate;
-    const showTTS = !!extra?.tts;
-
-    const showExtra = showModelTag || showTranslate || showTTS;
-
-    if (!showExtra) return;
-
     return (
-      <Flexbox gap={8} style={{ marginTop: 8 }}>
-        {showModelTag && (
-          <div>
-            <ModelTag model={extra?.fromModel as string} />
-          </div>
+      <Flexbox gap={8} style={{ marginTop: !!tools?.length ? 8 : 4 }}>
+        {content !== LOADING_FLAT && extra?.fromModel && (
+          <Usage
+            metadata={metadata || {}}
+            model={extra?.fromModel}
+            provider={extra.fromProvider!}
+          />
         )}
         <>
-          {extra?.tts && (
+          {!!extra?.tts && (
             <ExtraContainer>
               <TTS content={content} id={id} loading={loading} {...extra?.tts} />
             </ExtraContainer>
           )}
-          {extra?.translate && (
+          {!!extra?.translate && (
             <ExtraContainer>
               <Translate id={id} loading={loading} {...extra?.translate} />
             </ExtraContainer>