+++ src/libs/agent-runtime/utils/streams/openai.test.ts
@@ -348,6 +348,200 @@ describe('OpenAIStream', () => {
     ]);
   });
 
+  describe('token usage', () => {
+    it('should streaming token usage', async () => {
+      const data = [
+        {
+          id: 'chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          object: 'chat.completion.chunk',
+          created: 1741056525,
+          model: 'gpt-4o-mini-2024-07-18',
+          choices: [{ index: 0, delta: { role: 'assistant', content: '' } }],
+          service_tier: 'default',
+          system_fingerprint: 'fp_06737a9306',
+        },
+        {
+          id: 'chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          object: 'chat.completion.chunk',
+          created: 1741056525,
+          model: 'gpt-4o-mini-2024-07-18',
+          choices: [{ index: 0, delta: { content: '你好！' } }],
+          service_tier: 'default',
+          system_fingerprint: 'fp_06737a9306',
+        },
+        {
+          id: 'chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          object: 'chat.completion.chunk',
+          created: 1741056525,
+          model: 'gpt-4o-mini-2024-07-18',
+          choices: [{ index: 0, delta: {}, finish_reason: 'stop' }],
+          service_tier: 'default',
+          system_fingerprint: 'fp_06737a9306',
+        },
+        {
+          id: 'chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          object: 'chat.completion.chunk',
+          created: 1741056525,
+          model: 'gpt-4o-mini-2024-07-18',
+          choices: [],
+          service_tier: 'default',
+          system_fingerprint: 'fp_06737a9306',
+          usage: {
+            prompt_tokens: 1646,
+            completion_tokens: 11,
+            total_tokens: 1657,
+            prompt_tokens_details: { audio_tokens: 0, cached_tokens: 0 },
+            completion_tokens_details: {
+              accepted_prediction_tokens: 0,
+              audio_tokens: 0,
+              reasoning_tokens: 0,
+              rejected_prediction_tokens: 0,
+            },
+          },
+        },
+      ];
+
+      const mockOpenAIStream = new ReadableStream({
+        start(controller) {
+          data.forEach((chunk) => {
+            controller.enqueue(chunk);
+          });
+
+          controller.close();
+        },
+      });
+
+      const protocolStream = OpenAIStream(mockOpenAIStream);
+
+      const decoder = new TextDecoder();
+      const chunks = [];
+
+      // @ts-ignore
+      for await (const chunk of protocolStream) {
+        chunks.push(decoder.decode(chunk, { stream: true }));
+      }
+
+      expect(chunks).toEqual(
+        [
+          'id: chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          'event: text',
+          `data: ""\n`,
+          'id: chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          'event: text',
+          `data: "你好！"\n`,
+          'id: chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          'event: stop',
+          `data: "stop"\n`,
+          'id: chatcmpl-B7CcnaeK3jqWBMOhxg7SSKFwlk7dC',
+          'event: usage',
+          `data: {"inputCacheMissTokens":1646,"inputTextTokens":1646,"outputTextTokens":11,"totalInputTokens":1646,"totalOutputTokens":11,"totalTokens":1657}\n`,
+        ].map((i) => `${i}\n`),
+      );
+    });
+
+    it('should streaming litellm token usage', async () => {
+      const data = [
+        {
+          id: 'chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          created: 1741188058,
+          model: 'gpt-4o-mini',
+          object: 'chat.completion.chunk',
+          system_fingerprint: 'fp_06737a9306',
+          choices: [{ index: 0, delta: { content: ' #' } }],
+          stream_options: { include_usage: true },
+        },
+        {
+          id: 'chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          created: 1741188068,
+          model: 'gpt-4o-mini',
+          object: 'chat.completion.chunk',
+          system_fingerprint: 'fp_06737a9306',
+          choices: [{ index: 0, delta: { content: '.' } }],
+          stream_options: { include_usage: true },
+        },
+        {
+          id: 'chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          created: 1741188068,
+          model: 'gpt-4o-mini',
+          object: 'chat.completion.chunk',
+          system_fingerprint: 'fp_06737a9306',
+          choices: [{ finish_reason: 'stop', index: 0, delta: {} }],
+          stream_options: { include_usage: true },
+        },
+        {
+          id: 'chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          created: 1741188068,
+          model: 'gpt-4o-mini',
+          object: 'chat.completion.chunk',
+          system_fingerprint: 'fp_06737a9306',
+          choices: [{ index: 0, delta: {} }],
+          stream_options: { include_usage: true },
+        },
+        {
+          id: 'chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          created: 1741188068,
+          model: 'gpt-4o-mini',
+          object: 'chat.completion.chunk',
+          system_fingerprint: 'fp_06737a9306',
+          choices: [{ index: 0, delta: {} }],
+          stream_options: { include_usage: true },
+          usage: {
+            completion_tokens: 1720,
+            prompt_tokens: 1797,
+            total_tokens: 3517,
+            completion_tokens_details: {
+              accepted_prediction_tokens: 0,
+              audio_tokens: 0,
+              reasoning_tokens: 0,
+              rejected_prediction_tokens: 0,
+            },
+            prompt_tokens_details: { audio_tokens: 0, cached_tokens: 0 },
+          },
+        },
+      ];
+
+      const mockOpenAIStream = new ReadableStream({
+        start(controller) {
+          data.forEach((chunk) => {
+            controller.enqueue(chunk);
+          });
+
+          controller.close();
+        },
+      });
+
+      const protocolStream = OpenAIStream(mockOpenAIStream);
+
+      const decoder = new TextDecoder();
+      const chunks = [];
+
+      // @ts-ignore
+      for await (const chunk of protocolStream) {
+        chunks.push(decoder.decode(chunk, { stream: true }));
+      }
+
+      expect(chunks).toEqual(
+        [
+          'id: chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          'event: text',
+          `data: " #"\n`,
+          'id: chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          'event: text',
+          `data: "."\n`,
+          'id: chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          'event: stop',
+          `data: "stop"\n`,
+          'id: chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          'event: data',
+          `data: {"delta":{},"id":"chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5","index":0}\n`,
+          'id: chatcmpl-c1f6a6a6-fcf8-463a-96bf-cf634d3e98a5',
+          'event: usage',
+          `data: {"inputCacheMissTokens":1797,"inputTextTokens":1797,"outputTextTokens":1720,"totalInputTokens":1797,"totalOutputTokens":1720,"totalTokens":3517}\n`,
+        ].map((i) => `${i}\n`),
+      );
+    });
+  });
+
   describe('Tools Calling', () => {
     it('should handle OpenAI official tool calls', async () => {
       const mockOpenAIStream = new ReadableStream({
@@ -749,8 +943,8 @@ describe('OpenAIStream', () => {
           'event: text',
           `data: "帮助。"\n`,
           'id: 1',
-          'event: stop',
-          `data: "stop"\n`,
+          'event: usage',
+          `data: {"inputCacheMissTokens":6,"inputTextTokens":6,"outputReasoningTokens":70,"outputTextTokens":34,"totalInputTokens":6,"totalOutputTokens":104,"totalTokens":110}\n`,
         ].map((i) => `${i}\n`),
       );
     });
@@ -968,8 +1162,8 @@ describe('OpenAIStream', () => {
           'event: text',
           `data: "帮助。"\n`,
           'id: 1',
-          'event: stop',
-          `data: "stop"\n`,
+          'event: usage',
+          `data: {"inputCacheMissTokens":6,"inputTextTokens":6,"outputReasoningTokens":70,"outputTextTokens":34,"totalInputTokens":6,"totalOutputTokens":104,"totalTokens":110}\n`,
         ].map((i) => `${i}\n`),
       );
     });
@@ -1169,8 +1363,8 @@ describe('OpenAIStream', () => {
           'event: text',
           `data: "帮助。"\n`,
           'id: 1',
-          'event: stop',
-          `data: "stop"\n`,
+          'event: usage',
+          `data: {"inputCacheMissTokens":6,"inputTextTokens":6,"outputReasoningTokens":70,"outputTextTokens":34,"totalInputTokens":6,"totalOutputTokens":104,"totalTokens":110}\n`,
         ].map((i) => `${i}\n`),
       );
     });
@@ -1370,8 +1564,8 @@ describe('OpenAIStream', () => {
           'event: text',
           `data: "帮助。"\n`,
           'id: 1',
-          'event: stop',
-          `data: "stop"\n`,
+          'event: usage',
+          `data: {"inputCacheMissTokens":6,"inputTextTokens":6,"outputReasoningTokens":70,"outputTextTokens":34,"totalInputTokens":6,"totalOutputTokens":104,"totalTokens":110}\n`,
         ].map((i) => `${i}\n`),
       );
     });
@@ -1571,8 +1765,8 @@ describe('OpenAIStream', () => {
           'event: text',
           `data: "帮助。"\n`,
           'id: 1',
-          'event: stop',
-          `data: "stop"\n`,
+          'event: usage',
+          `data: {"inputCacheMissTokens":6,"inputTextTokens":6,"outputReasoningTokens":70,"outputTextTokens":34,"totalInputTokens":6,"totalOutputTokens":104,"totalTokens":110}\n`,
         ].map((i) => `${i}\n`),
       );
     });