+++ src/libs/agent-runtime/utils/streams/protocol.ts
@@ -1,4 +1,5 @@
 import { ChatStreamCallbacks } from '@/libs/agent-runtime';
+import { ModelTokensUsage } from '@/types/message';
 
 import { AgentRuntimeErrorType } from '../../error';
 
@@ -9,9 +10,10 @@ export interface StreamContext {
   id: string;
   /**
    * As pplx citations is in every chunk, but we only need to return it once
-   * this flag is used to check if the pplx citation is returned,and then not return it again
+   * this flag is used to check if the pplx citation is returned,and then not return it again.
+   * Same as Hunyuan and Wenxin
    */
-  returnedPplxCitation?: boolean;
+  returnedCitation?: boolean;
   thinking?: {
     id: string;
     name: string;
@@ -22,6 +24,7 @@ export interface StreamContext {
     name: string;
   };
   toolIndex?: number;
+  usage?: ModelTokensUsage;
 }
 
 export interface StreamProtocolChunk {
@@ -43,6 +46,8 @@ export interface StreamProtocolChunk {
     | 'stop'
     // Error
     | 'error'
+    // token usage
+    | 'usage'
     // unknown data result
     | 'data';
 }